Fix edge_attr dtype to torch.long, add resume and multi-run support
- utils.py: cast edge_attr to torch.long (matching authors' preprocessed
  data which uses int64 for all edge features)
- train.py: add --resume flag with checkpoint save/load, add run_suffix
  for multi-run naming, tqdm progress bars
- README.md: add non-uv setup instructions via setup_env.sh

+++ utils.py
@@ -184,6 +184,5 @@ def get_gnn_features(protein, ligand, threshhold=5):
         edge_features_list.append(edge_feature)
 
     edge_index = torch.tensor(np.array(edges_list).T, dtype=torch.long)
-    edge_attr = torch.tensor(np.array(edge_features_list),
-                             dtype=torch.float)
+    edge_attr = torch.tensor(np.array(edge_features_list), dtype=torch.long)
     return x, edge_index, edge_attr
